Week_03/swebench_proposed/requirements.txt
# Minimal dependencies for SWE-bench proposed replication
openai>=1.0.0
python-dotenv>=1.0.0
datasets>=2.14.0
gitpython>=3.1.0
tqdm>=4.65.0
requests>=2.31.0
pytest>=8.0.0
pyyaml>=6.0
